Remove open client class
